Manually specify some table column widths
Turning on wrapping in the theme leaves decisions on column widths
up to the Sphinx formatter. It does an OK job sometimes, other times
the tables are imbalanced.

Column widths can be manually specified by using the explicit
`.. table:` directive, using its `:widths:` option.

(Table code itself is the same, but has to be nested inside the
`.. table:` directive, requiring the entire table block be
indented five spaces. Not ideal, but it's what works.)

Width syntax is... complex. If specified for all columns out of
100, then they appear to effectively be percentages. Widths can
be left out for some columns (on the right of the table), and
they will be defaulted. But that computation involves some sort
of Sphinx magic.

If we have a three-column table, `:widths: 5 25` (with the third
column unspecified) will _not_ be laid out the same as
`:widths: 5 25 70`.

I've allowed Sphinx to default column widths where possible, but
when things got weird or unpredictable I specified explicit widths
for all columns.