Re-work communication with child windows
This does away with the main window's (expensive) event filter,
which was previously used to communicate state to tutorial popups.

Instead it reimplements QMainWindow's event processing methods
`moveEvent()`, `resizeEvent()`, `showEvent()`, and `hideEvent()`
to propagate certain events to ALL floating QDockWidgets. This
way, if (for example) Properties is floating and OpenShot is
minimized, the floating dock will minimize with it, and it will
restore again when the application is unminimized.

By turning the tutorial popups into floating QDockWidgets (separate
commit), they will also benefit from this change.

Move and Resize events on the main window trigger a call to the
tutorial manager's `re_position_dialog()`, which updates the
placement of the tutorial popup (if visible).