Convert tutorial popups into floating QDockWidget
This commit radically alters the structure of the tutorial system,
based on some research and a half-formed suggestion I saw on some
StackOverflow question related to Qt window management.

It uses the same floating, non-dockable QDockWidget (initially hidden)
to display all of the popups. The widget contained within that dock
is replaced for each tutorial step. The dock has no titlebar or frame,
and should appear identical to the previous tutorial popup windows.

Because the tutorials are held in a QDockWidget, Qt automatically
takes care of keeping the tutorial popups in front of the main
application window.

The event filter is once again done away with, and calls to the
`show()` and `raise_()` methods are tightly managed to prevent
unnecessary interface updates. My hope is that this will dramatically
reduce the flicker and GUI unresponsiveness that some users have
encountered.

This is not quite complete, and requires extensive testing on
platforms other than Linux especially. I'm making this WIP commit
in order to test the current code under Windows.

Currently unresolved:

* Qt will keep the tutorial boxes above the QMainWindow, but I'm
having trouble keeping the tutorial popups from ending up behind
OTHER floating QDockWidgets, especially after a minimize/restore
cycle.