Don't skip loading system locale
PR #1332 sought to fix loading of English translations in non-English
locales (see #1256) by removing a hardcoded `en_US` bypass in the
translation file loading. In doing so, it inadvertently broke loading
of translation files that match the system locale.

This removes the entire "skip loading" clause, because we can never
assume that any translation shouldn't be loaded. English translations
may need to be loaded and applied, if OpenShot's language is set to
English in non-English locales. In contrast, we still need to load our
OpenShot translations for the local language even if the selected
language is the same as the system locale.