Set up QResource index for translations
This completely reworks the on-disk model for translation files,
doing away with the `lang_code/LC_MESSAGES/` path hierarchy entirely
in favor of a flat directory structure containing files named with
the pattern "OpenShot.lang_code.qm". This allows us to take advantage
of QTranslator's fallback code that will, for instance, automatically
load `OpenShot.fr.qm` when `OpenShot.fr_ZZ.qm` is not found.

The translation files are also indexed in a QResource file named
`openshot_lang.qrc`, which is compiled into `openshot_lang.py`. This
allows the translations to be loaded from the resource path ":/locale".

The translations are stored in the `language/` subdirectory, instead
of `locale/`, because `from locale import openshot_lang` conflicts
with the Python module named `locale`.