Restore the special-casing override for en_US
This change _properly_ special-cases for `en_US` as the preference
locale, by completely **replacing** the list of possible translations
to load. This ensures that no language other than `en_US` is pulled
in by the QTranslator. Since we _know_ that we have full support
for `en_US` and always will, there's no need to keep a list of other
possible languages from the locale/environment to fall back on.

Add OpenShot's language preference to the locale-environment logging.