Gracefully handle missing `openshot_lang.py`
If the compiled language resource file `src/language/openshot_lang.py`
is not found on import, the translation code will now gracefully fall
back to scanning `src/language/` as a normal filesystem directory, and
automatically pick up all of the `OpenShot.*.qm` files it contains.
So, this actually reduces the lock-in on the compiled resource file
considerably.

Note that it's still all-or-nothing: If the compiled language resources
are found, OpenShot will not look at the directory. So, to add any new
or updated translation files, they should be added to the directory, and
then `openshot_lang.py` should be **deleted**. (Or, preferably, edited
to pick up the new file and then regenerated.)