Add debug logging level to logfile (only)
This commit enables something I've been working towards for some
time now: The ability to configure OpenShot's console output and
logfile output to different logging levels, with potentially _more_
messages being logged to the logfile than are shown on the console.

As a result, it brings all of the following:
* Log messages of level INFO or higher are written to the console
  log
* By default, initially the same messages are written to the
  logfile as well
* The `debug-mode` preference now affects openshot-qt's own logging,
  as well as the libopenshot proxy logging. When `debug-mode` is
  switched on, the `openshot-qt.log` file will receive all messages
  of level DEBUG or higher.
* Log messages proxied from libopenshot are now logged at level
  DEBUG, instead of INFO, so that they will be written to the
  log file (only if `debug-mode` is on), but will not be printed
  to the console output.

This opens up the possibility of adding a second ZeroMQ logging
channel for messages at a _higher_ severity than debug, which
would always be active rather than being gated by the `debug-mode`
switch. That priority logging channel could be used for messages
which openshot should always log, such as error messages. This
would be a possible means of addressing #1450.