Install icons to freedesktop paths from setup.py
The `setup.py` installer in the repo root directory has only
been installing the SVG icon file, and into `/usr/share/pixmaps/`
which is a path only used by the Debian menu system[1] (not by
XDG, as the file claims.)

XDG uses Freedesktop's icon theme specification[2], which instructs
application developers to install their icon files in the theme
directories: `$prefix/share/icons/hicolor/*/apps` (sorted by size).

This change installs the existing `openshot-qt.svg` scalable icon
and `openshot-qt.png` (64x64) icon, as well as the new 256x256 and
512x512 PNGs, into their appropriate places in the icon theme
hierarchy.

Again, this is something we've been taking care of "manually" on
the packaging end. I'd prefer to offload it to `setup.py`, which
we're running anyway as part of the package-build process, so that
everyone building the package can benefit from it.

[1]: https://www.debian.org/doc/packaging-manuals/menu.html/ch3.html#s3.7
[2]: https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-latest.html#install_icons