Fix the Supporters-confused-for-contributors bug
Commit 4b4172d111 updated the contributors.json file with the data
that had to go to supporters.json file, cauing the 'credits' in
'About OpenShot' window to list the supporters (the financial back-
ers) as the developers. This commit renames what was in the previo-
us commit contributors.json into supporters.json and brings back
contributors.json file from the commit immediately preceding the
aforementioned faulty commit.