Default to three tracks for new projects
Five tracks is excessive as a starting configuration. Few users will
ever create projects that complex, and with the current default
window geometry not even able to fit _two_ tracks, it's just a waste
of space. This reduces the default track count to three for both new
projects and the debugging interface.