Change order of tabbing items in export dialog (#2782)
* Change order of tabbing items in export dialog

The order of tabbing with keyboard depends on UI creation.

* "Start Frame" (1)
* "End Frame" (2) - was created first

Thus, when field that lies in UI upper (the "Start Frame" field),
is created after the lower item (the "End Frame" field) the forward
tabbing order will be "End Frame" (2) >>> "Start Frame" (1)
instead of "Start Frame" (1) >>> "End Frame" (2)

Now items creation order is follow to its real layout:

* "Start Frame" (1) - now created first
* "End Frame" (2)

* Change order of some UI elements creation

Affects only tabbing order with the keyboard.
The UI items position is preserved.

* Change order of tabbing items in export dialog

This only affects tabbing order of the Simple tab of the export dialog
window. New order - from top to bottom, from left to right.

* Use Qt tabstops to manage tabbing order of items

In Export dialog of OpenShot use tabstops to control tabbing order of
UI items. This is implemented for convenience and easier managment of
the code in future (to not depend on items creation order).

Co-authored-by: Frank Dana <ferdnyc@users.noreply.github.com>