Added the ability to automatically scroll the timeline to points of interest.  Specifically, the user can center the timeline on the current playhead position.  Additionally, when the playhead is moved to a marker or the start or end of the timeline, the timeline is scrolled to match.
Added the function "centerOnTime" to controllers.js to allow the timeline to be centered on an arbitrary time.
Added the function "centerOnTime" to timeline_webview.py to call the "centerOnTime" function in controllers.js.
Added the action "actionCenterOnPlayhead" to main-window.ui and added it to the toolbox between "actionPreviousMarker" and "actionNextMarker".
Added the function "actionCenterOnPlayhead" to main_window.py to center the timeline on the current playhead position.
Modified main_window.py to automatically scroll the timeline when the playhead is moved to a marker or the start or end of the timeline.