Added the ability to center the timeline view at an arbitrary time.  Currently, this can be used by the user to automatically jump to the current playhead position rather than manually searching for it.  The timeline also recenters itself when jumping to a marker or to the beginning or end of the timeline.
Modified "main-window.ui" to add the "actionCenterOnPlayhead" action.
Modified "_default.settings" to add a new hotkey (Ctrl+Up) for the "ationCenterOnPlayhead" action.
Modified "controller.js" to add the "centerOnTime" function to perform the actual work of scrolling the timeline.
Modified "timeline_webview.py" to add the "centerOnTime" function as an intermediary for the function in "controller.js".
Modified "main_window.py" to add the "centerOnPlayhead_triggered" function to implement the action behavior.  Also added code to automatically recenter when jumping to the beginning or end of the timeline or moving to a marker.