Fix render progress display in Animated Titles editor (#2964)
* Animated titles: Update progress for Blender 2.80

Blender 2.80 doesn't output "Part" messages anymore, so the frame
regexp was always failing, and the progress bar never updated.

This switches to detecting the final "Sce:" message for the scene, and
using it to update the progress bar.

* Animated Titles: remove unused close_window()

* Animated titles: Remove more slot methods

`blender_listview.py` contained a lot of slot methods of the form:
```py
def onRenderFinish(self):
    self.render_finished()
```
There's no real point to that, as `render_finished()` can simply be
decorated as a slot and connected directly. So, we do that instead.

* blender_listview: Remove codecs module

`codecs.open()` is documented as a wrapper around `open()` that uses
`errors='strict'` by default, so we can just do that ourselves.