Don't reformat project file names
Both the code to set the main window title and the Export dialog's
generation of the initial base export filename were processing the
user's assigned project file name, replacing `_` and `-` characters
with spaces and altering the capitalization. That's simply rude, and
there's no reason to mess with the user's file naming.

If they want spaces in the name, they'll insert them. If they want
their project file to be named `MY_OPENSHOT_PROJECT.osp`, then we
should default to `MY_OPENSHOT_PROJECT.mp4` as the export filename.