Fix the updating of the type of previous point
The interpolation type of the previous point shouldn't be changed when
interpolation coefficients added.

This fixes issue when current and previous segment both modifies its
interpolation type, when only current was modified.

In libopenshot the closest_point_x is defines interpolation for the
segment, while previous_point_x only stores interpolation coefficients
(details) that may be needed to calculate the value for the
closest_point_x segment of the curve.

Co-authored-by: Stephane Chauveau <schauveau@users.noreply.github.com>