Docs: Add Sphinx directive for youtube embeds
Add a simple directive `.. youtube`, implemented as a local sphinx
extension, that takes one argument — either:
  - A YouTube video ID (e.g. `aqz-KE-bpKQ`)
  - A YouTube URL (e.g. `https://www.youtube.com/watch?v=aqz-KE-bpKQ`)

In either case, when generating HTML docs the directive will be
replaced with an HTML embed of the video content.

Valid forms for the URL include at least:
- https://www.youtube.com/watch?v=YOUTUBE_ID
- https://www.youtube.com/embed/YOUTUBE_ID
- https://youtu.be/YOUTUBE_ID

Features like start time (`?t=45`) are currently not supported.