Improve Documentation
This is a smallish attempt to improve openshot's sparse documentation
(issue #2989).
The getting started page only contains an installation description, so name
it so. Improve the quick tutorial somewhat (at least according to my taste)
but still keep it quick. Improve the start of the "Files" pages somewhat.

Make the import files table more consistent by nameing the method in the
first column and actually describing them in a full senctence (interpunction)
in the 2nd column. Add the actual toolbar button icon image to the table.

Similar things for the slice clips table on the clips page.