Suggestions from rexdk included
One example among many still in the original table: "Curve representing the crop height of the clip in percent (0 to 1)" - that is simply not true - the value is not in %, it is a fraction. 1 doesn't represent 1%, it represents 100%. And it is not a curve - it is a single number.\

You've kindly taken my outline box - but because it is not in the flow of the calculation of properties it now isn't defined correctly. It only works if it it defined at the correct moment of calculation.

There is a serious bug in the Time property - it does not work; it's likely that my update to this property on my cms has missed you.. My view is that the user guide should be honest about bugs - if something doesn't work as expected you can save an awful lot of user pain. I'm aware that OpenShot has rather a lot of bugs (I've found 7 or 8 so far, and a couple more I haven't been able to pin down) - but that's even more reason to help the user :)