classes/waveform.py: don't process every frame
Based on the use of "seconds" in the local variable names, I gather that
the intention of this function is to get a few audio samples from each
second of data.  However, what it was actually doing is getting that
many samples from each *frame* of data, which is way more data,
resulting in the process taking an unreasonably long time.