"Copy Keyframes > All" includes "Crop *" settings
(Feel free to ignore this PR or tell me i'm out of my mind)

But this may help someone who maybe expecting the Copy Keyframes > All feature to include the Crop keyframes (which work together with scaling X/Y and location X/Y to "frame" a video nicely).

Added data elements in my proposal:
-`crop_x`
-`crop_y`
-`crop_height`
-`crop_width`

Tested it locally and seems to work. Not sure if there is a catch. Again, please just tell me if there is.

Hope this helps