Move XML/EDL listings to appendix, wrap for display
- The XML listing is too long and unwieldy for a mid-document
  embed
- It's also far too wide
- With the XML relocated, the EDL listing sticks out as well.
  By removing both of them, that chapter becomes one comfy
  page in the PDF.

To address all of these problems, the listings are relegated to
an "Appendix A" at the end of the manual. References in both
sections link directly to the corresponding listings.

The appendix is, specifically, `appendix_a.md`, a MarkDown
file. (It was easier to edit and activate code-fenced
syntax highlighting.) Sphinx supports MarkDown handily, it's
just a matter of activating it and installing a parser.

The MySt processor (a Sphinx-compatible MarkDown parser) is
added to the requirements.txt as an additional docs dependency.

The new file contains a painstakingly hand-wrapped reformatting
of the XML data so it no longer overflows its bounding boxes
and causes major badness in the PDF output.

Other than being reflowed, there are no changes to the XML.