title_editor: fix multi-lines titles update error if missing line 1
Example: "Standard 3" title:

* if line 1 and line 3 are missing, then line 2 is immutable (can not update).

* if line 1 is missing, then update to line 2 is lost and update to line 3 becomes line 2.

Note that if line 3 is missing, then update to line 1 and line 2 are working fine.

The fix is to always display all the lines, even if they are empty,
which fixes all the update problems above.

However, those lines that are empty, they are not updatable, even if the
field is writable.